+++ brickrail-gui/layout/block/layout_logical_block.gd
@@ -12,6 +12,8 @@ var selected = false
 var _hover = false
 var can_stop = true
 var can_flip = true
+var random_target = true
+var wait_time = 4.0
 
 var LayoutBlockInspector = preload("res://layout/block/layout_block_inspector.tscn")
 
@@ -35,6 +37,16 @@ func set_section(p_section):
 	section = p_section
 	section.connect("sensor_changed", self, "_on_section_sensor_changed")
 	find_sensors()
+
+func set_random_target(value):
+	if value != random_target:
+		LayoutInfo.set_layout_changed(true)
+	random_target = value
+
+func set_wait_time(value):
+	if value != wait_time:
+		LayoutInfo.set_layout_changed(true)
+	wait_time = float(value)
 
 func _on_section_sensor_changed():
 	find_sensors()